decoder: Extract Any ConditionalExpr logic into one file
