+++ decoder/expr_any_ref_origins.go
@@ -283,42 +283,3 @@ func (a Any) refOriginsForTemplateExpr(ctx context.Context, allowSelfRefs bool) 
 
 	return origins, false
 }
-
-func (a Any) refOriginsForConditionalExpr(ctx context.Context, allowSelfRefs bool) (reference.Origins, bool) {
-	origins := make(reference.Origins, 0)
-
-	// There is currently no way of decoding conditional expressions in JSON
-	// so we just collect them using the fallback logic assuming "any"
-	// constraint and focus on collecting expressions in HCL with more
-	// accurate constraints below.
-
-	switch eType := a.expr.(type) {
-	case *hclsyntax.ConditionalExpr:
-		condExpr := newExpression(a.pathCtx, eType.Condition, schema.AnyExpression{
-			OfType: cty.Bool,
-		})
-		if expr, ok := condExpr.(ReferenceOriginsExpression); ok {
-			origins = append(origins, expr.ReferenceOrigins(ctx, allowSelfRefs)...)
-		}
-
-		trueExpr := newExpression(a.pathCtx, eType.TrueResult, schema.AnyExpression{
-			OfType:                  a.cons.OfType,
-			SkipLiteralComplexTypes: a.cons.SkipLiteralComplexTypes,
-		})
-		if expr, ok := trueExpr.(ReferenceOriginsExpression); ok {
-			origins = append(origins, expr.ReferenceOrigins(ctx, allowSelfRefs)...)
-		}
-
-		falseExpr := newExpression(a.pathCtx, eType.FalseResult, schema.AnyExpression{
-			OfType:                  a.cons.OfType,
-			SkipLiteralComplexTypes: a.cons.SkipLiteralComplexTypes,
-		})
-		if expr, ok := falseExpr.(ReferenceOriginsExpression); ok {
-			origins = append(origins, expr.ReferenceOrigins(ctx, allowSelfRefs)...)
-		}
-
-		return origins, true
-	}
-
-	return origins, false
-}
